% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{generate()}
\alias{generate()}
\title{Generate}
\arguments{
\item{size}{length of output.}

\item{dict}{choose one of c("url", "numbers", "lowercase", "uppercase", "nolookalikes") or give acceptable strings set as character.}

\item{init.locales}{vocabulary sets of obscene words. defaut sets are c("en", "ja").}
}
\value{
strings output.
}
\description{
Generate
}
