% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRandomValues}
\alias{getRandomValues}
\title{Pseudo window.crypto.getRandomValues() function}
\usage{
getRandomValues(seed = NULL, size = 64L)
}
\arguments{
\item{seed}{string.}

\item{size}{integer (size of output).}
}
\value{
list of integers.
}
\description{
The original ai/nanoid generates random strings
with window.crypto.getRandomValues() on client side environment,
however, the V8 environment has no crypto object.
Because of this, this package replaces that function with
an original random value generator.
}
\seealso{
\url{https://developer.mozilla.org/ja/docs/Web/API/Window/crypto}
}
