% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Locales}
\alias{Locales}
\title{List of vocabularies dict}
\source{
\url{https://github.com/LDNOOBW/naughty-words-js}
}
\description{
The source of vocabulary comes from LDNOOBW/naughty-words-js.
You can pass locale `Code` of this list to NanoID$format() and NanoID$generate().
}
