% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{NanoID$format()}
\alias{NanoID$format()}
\title{Format}
\arguments{
\item{size}{length of output.}

\item{dict}{choose one of c("url", "numbers", "lowercase", "uppercase", "nolookalikes") or give acceptable strings set as character.}

\item{init.locales}{vocabulary sets of words. defaut sets are c("en", "ja").}

\item{use_func}{function name that will be used for generating buffer array.}
}
\value{
strings output.
}
\description{
Customizable generator that you can specify your own buffer-creation method.
You can specify any R function that recieves `size` as its argument
and returns a buffer array of wihch length is `size`.
}
