% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{NanoID$nonsecure()}
\alias{NanoID$nonsecure()}
\title{Nonsecure}
\arguments{
\item{size}{length of output.}

\item{init.locales}{vocabulary sets of obscene words. defaut sets are c("en", "ja").}
}
\value{
strings output.
}
\description{
Faster but non-secure version of NanoID generator (bigger collision probability).
}
