% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoidGood.R
\name{NanoID-class}
\alias{NanoID-class}
\alias{NanoIdGood}
\title{NanoID class}
\description{
R6 class definition of NanoID-class.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ctx}}{V8 context.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NanoIdGood$new()}}
\item \href{#method-format}{\code{NanoIdGood$format()}}
\item \href{#method-generate}{\code{NanoIdGood$generate()}}
\item \href{#method-nonsecure}{\code{NanoIdGood$nonsecure()}}
\item \href{#method-clone}{\code{NanoIdGood$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NanoIdGood$new(ctx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ctx}}{V8 context.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
Format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NanoIdGood$format(
  size = 21L,
  dict = c("url", "numbers", "lowercase", "uppercase", "nolookalikes"),
  init.locales = c("en", "ja"),
  use_func = "randombytes"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{length of output.}

\item{\code{dict}}{choose one of c("url", "numbers", "lowercase", "uppercase", "nolookalikes")
or give acceptable strings set as character.}

\item{\code{init.locales}}{vocabulary sets of words.
default sets are c("en", "ja").}

\item{\code{use_func}}{function name that will be used for generating bytes array.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Customizable generator that you can specify your own bytes-array-creation method.
You can specify any R function that receives `size` as its argument
and returns a bytes-array of which length is `size`.
}

\subsection{Returns}{
strings output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate"></a>}}
\if{latex}{\out{\hypertarget{method-generate}{}}}
\subsection{Method \code{generate()}}{
Generate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NanoIdGood$generate(
  size = 21L,
  dict = c("url", "numbers", "lowercase", "uppercase", "nolookalikes"),
  init.locales = c("en", "ja")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{length of output.}

\item{\code{dict}}{choose one of c("url", "numbers", "lowercase", "uppercase", "nolookalikes")
or give acceptable strings set as a character.}

\item{\code{init.locales}}{vocabulary sets of obscene words.
default sets are c("en", "ja").}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
strings output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nonsecure"></a>}}
\if{latex}{\out{\hypertarget{method-nonsecure}{}}}
\subsection{Method \code{nonsecure()}}{
Nonsecure
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NanoIdGood$nonsecure(size = 21L, init.locales = c("en", "ja"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{length of output.}

\item{\code{init.locales}}{vocabulary sets of obscene words.
default sets are c("en", "ja").}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Faster but non-secure version of NanoID generator
(bigger collision probability).
}

\subsection{Returns}{
strings output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NanoIdGood$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
