% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{nanoid}
\alias{nanoid}
\title{Get ready to use NanoID}
\usage{
nanoid(ctx = NULL)
}
\arguments{
\item{ctx}{Your own V8 context if any.}
}
\value{
R6 class object having methods below.
\itemize{
  \item format()
  \item generate()
  \item nonsecure()
  \item getContext()
}
}
\description{
Get ready to use NanoID
}
\seealso{
\url{https://github.com/ai/nanoid/blob/master/README.md}
}
