% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoid.R
\name{nanoid_impl}
\alias{nanoid_impl}
\title{NanoID generator}
\usage{
nanoid_impl(size, alphabet, func)
}
\arguments{
\item{alphabet}{Character vector.}

\item{func}{Function.}

\item{step}{Integer.}
}
\value{
Character vector.
}
\description{
An R implementation of NanoID.
}
\keyword{internal}
