% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoidr-package.R
\docType{package}
\name{nanoidr-package}
\alias{nanoidr}
\alias{nanoidr-package}
\title{NanoID for R}
\description{
An R port of NanoID that is a tiny, secure,
URL-friendly, unique string ID generator for JavaScript.

The original NanoID library generates random strings using crypto API,
which is replaced with openssl::rand_bytes in nanoidr package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/nanoidr}
  \item \url{https://paithiov909.github.io/nanoidr/}
  \item Report bugs at \url{https://github.com/paithiov909/nanoidr/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
