% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R, R/nanoidr.R
\docType{package}
\name{nanoidr}
\alias{nanoidr}
\alias{nanoidr-package}
\title{Ready to Use Nano ID}
\usage{
nanoidr(ctx = NULL)
}
\arguments{
\item{ctx}{Your own V8 context if any.}
}
\value{
R6 class object having methods below.
\itemize{
  \item format()
  \item generate()
  \item nonsecure()
}
}
\description{
Ready to Use Nano ID

Nano ID on V8
}
