% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{no_lookalikes}
\alias{no_lookalikes}
\title{Predefined character sets}
\usage{
no_lookalikes(pattern = c("nolookalikes", "nolookalikessafe"))
}
\arguments{
\item{pattern}{Choice one of `nolookalikes` or `nolookalikessafe`.}
}
\value{
Character vectors.
}
\description{
Return a predefined character set ported of the 'nanoid-dictionary'.
This function supports 2 patterns; "nolookalikes" or "nolookalikesSafe".
}
\examples{
no_lookalikes("nolookalikes")
no_lookalikes("nolookalikessafe")
}
