% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{randombytes}
\alias{randombytes}
\title{An example of original random bytes generator for using in fomrat()}
\usage{
randombytes(size)
}
\arguments{
\item{size}{integer passed to `sample(1:256, size, replace = TRUE)`}
}
\value{
function.
}
\description{
An example of original random bytes generator for using in fomrat()
}
