% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{randstr}
\alias{randstr}
\title{Wrapper function of stringi::stri_rand_strings()}
\usage{
randstr(size)
}
\arguments{
\item{size}{Length of random string.}
}
\value{
function.
}
\description{
Wrapper function of stringi::stri_rand_strings()
}
