% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRandomValues}
\alias{getRandomValues}
\title{Partial func stringi::stri_rand_strings()}
\usage{
getRandomValues(seed)
}
\value{
function.
}
\description{
Creates a function that invokes stringi::stri_rand_strings()
with partials prepended to the arguments it receives.
}
