% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRandomValues}
\alias{getRandomValues}
\title{Pseudo crypto.getRandomValues Web API}
\usage{
getRandomValues(size = 64L, seed = 1234L, fix.seed = FALSE)
}
\arguments{
\item{size}{integer (size of output).}

\item{seed}{only when fix.seed is true, will be used for set.seed(). Or not, will be used just for dqrng::dqset.seed().}

\item{fix.seed}{default is false.}
}
\value{
list of buffer.
}
\description{
The original NanoID library generates random strings
with window.crypto.getRandomValues on client side environment,
however, the V8 environment provides no crypto object.
Because of this, {nanoidr} package replaces that function with
a simple wrapper of openssl::rand_bytes().
}
\seealso{
\url{https://developer.mozilla.org/ja/docs/Web/API/Window/crypto}
}
