% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Locales}
\alias{Locales}
\title{List of built in dictionaries}
\source{
\url{https://github.com/LDNOOBW/naughty-words-js}
}
\description{
The source of dictionaries comes from LDNOOBW/naughty-words-js.
You can pass locale `Code` strings of this list to NanoID$format() and NanoID$generate().
}
