% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoid.R
\name{nanoid}
\alias{nanoid}
\title{Generate NanoID}
\usage{
nanoid(size = 21L, alphabet = NULL, func = NULL)
}
\arguments{
\item{size}{Integer.}

\item{alphabet}{Character vector or NULL.}

\item{func}{Function or NULL.}
}
\value{
String.
}
\description{
Generate NanoID
}
