% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoidGood.R
\name{nanoid}
\alias{nanoid}
\title{Get ready to use NanoID}
\usage{
nanoid(ctx = NULL)
}
\arguments{
\item{ctx}{your own V8 context if any.}
}
\value{
R6 class object
}
\description{
Initilaize an R6 class object that has context for using NanoID
(y-gagar1n/nanoid-good) functions.
}
\details{
The class object has methods below.
\itemize{
  \item format()
  \item generate()
  \item nonsecure()
}

Especially, `format()` and `generate()` methods check every generated ID
through a vocabulary of obscene words specified in `dict` argument.
If any match is found, then ID is generated again and again
until it gets clean ID which is returned.
}
\seealso{
\url{https://github.com/ai/nanoid/blob/master/README.md}

\url{https://github.com/y-gagar1n/nanoid-good/blob/master/README.md}
}
