% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoidAlt.R
\name{nanoidAlt}
\alias{nanoidAlt}
\title{Yet another R implementation of NanoID}
\usage{
nanoidAlt(size = 21L)
}
\arguments{
\item{size}{expected output size.}
}
\value{
strings output.
}
\description{
Yet another R implementation of NanoID
}
