% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoid.R
\docType{package}
\name{nanoidr}
\alias{nanoidr}
\alias{nanoidr-package}
\title{NanoID on V8}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

NanoID (an ID generator for JavaScript) on V8 for R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/nanoidr}
  \item Report bugs at \url{https://github.com/paithiov909/nanoidr/issues}
}

}
\author{
\strong{Maintainer}: Kato Akiru \email{paithiov909@gmail.com}

}
\keyword{internal}
